test: consolidate low-risk test suites
