+++ tests/unit/storage/test_storage_registry.py
@@ -1,6 +1,8 @@
 # pyright: reportPrivateImportUsage = false, reportPrivateUsage = false
 """Unit tests for StorageRegistry."""
 
+import importlib.util
+import types
 from pathlib import Path
 
 import pytest
@@ -9,6 +11,16 @@ from sqlspec.exceptions import ImproperConfigurationError, MissingDependencyErro
 from sqlspec.storage.registry import StorageRegistry
 from sqlspec.typing import FSSPEC_INSTALLED, OBSTORE_INSTALLED
 from sqlspec.utils.type_guards import is_local_path
+
+
+def _load_registry_source_module() -> "types.ModuleType":
+    module_path = Path(__file__).resolve().parents[3] / "sqlspec" / "storage" / "registry.py"
+    spec = importlib.util.spec_from_file_location("storage_registry_source_tests", module_path)
+    assert spec is not None
+    assert spec.loader is not None
+    module = importlib.util.module_from_spec(spec)
+    spec.loader.exec_module(module)
+    return module
 
 
 def test_is_local_path_type_guard() -> None:
@@ -181,6 +193,18 @@ def test_clear_cache(tmp_path: Path) -> None:
     backend1 = registry.get(f"file://{tmp_path}")
     registry.clear_cache(f"file://{tmp_path}")
     backend2 = registry.get(f"file://{tmp_path}")
+
+    assert backend1 is not backend2
+
+
+def test_source_registry_clear_cache_removes_parameterized_uri_entries(tmp_path: Path) -> None:
+    """Source implementation should evict backend-override cache entries by base URI."""
+    module = _load_registry_source_module()
+    registry = module.StorageRegistry()
+
+    backend1 = registry.get(f"file://{tmp_path}", backend="local")
+    registry.clear_cache(f"file://{tmp_path}")
+    backend2 = registry.get(f"file://{tmp_path}", backend="local")
 
     assert backend1 is not backend2
